refactor: split configeditor/update_records.go one file per mode
529 lines holding the key handling for all four record-editor modes, split
the way the file's own section markers already divided it:

  update_records_field.go    222  field editing: open, keys, apply, scroll
  update_records.go          158  record list browsing + list scroll clamp
  update_records_edit.go     103  moving between a record's fields
  update_records_reorder.go   63  moving a record within its list

Pure move: no declaration body changed. Verified with a go/parser pairwise
diff of every top-level declaration (12 old, 12 new, 0 missing, 0 mismatched,
0 extra) and a comment multiset diff.

The five "// --- X Mode ---" section markers are the only comments removed;
each file now carries a header describing what it holds, which is what those
markers were doing inside one file. No test file touched.
